Rename module influence.dask.influence_model to influence.dask.influence_calculator
