+++ src/pydvl/influence/dask/__init__.py
@@ -1,1 +1,1 @@
-from .influence_model import DaskInfluenceCalculator
+from .influence_calculator import DaskInfluenceCalculator
